Move default oauth responses to the core library
