+++ backend/t4cclient/core/oauth/__init__.py
@@ -1,0 +1,3 @@
+# Copyright DB Netz AG and the capella-collab-manager contributors
+# SPDX-License-Identifier: Apache-2.0
+
