+++ backend/t4cclient/extensions/backups/ease/routes.py
@@ -15,7 +15,7 @@ from t4cclient.core.authentication.jwt_bearer import JWTBearer
 from t4cclient.core.database import get_db
 from t4cclient.sessions.operators import OPERATOR
 from t4cclient.extensions.modelsources import git, t4c
-from t4cclient.routes.open_api_configuration import AUTHENTICATION_RESPONSES
+from t4cclient.core.oauth.responses import AUTHENTICATION_RESPONSES
 
 from . import crud, helper, models
