+++ backend/t4cclient/extensions/backups/jenkins/routes.py
@@ -16,7 +16,7 @@ from t4cclient.core.authentication.database.jenkins import \
 from t4cclient.core.authentication.jwt_bearer import JWTBearer
 from t4cclient.core.database import get_db
 from t4cclient.extensions.backups import jenkins
-from t4cclient.routes.open_api_configuration import AUTHENTICATION_RESPONSES
+from t4cclient.core.oauth.responses import AUTHENTICATION_RESPONSES
 
 from . import crud as jenkins_database
