+++ backend/t4cclient/extensions/modelsources/t4c/routes.py
@@ -11,7 +11,7 @@ from requests import Session
 from t4cclient.core.authentication.database import (verify_admin,
                                                     verify_repository_role)
 from t4cclient.core.authentication.jwt_bearer import JWTBearer
-from t4cclient.routes.open_api_configuration import AUTHENTICATION_RESPONSES
+from t4cclient.core.oauth.responses import AUTHENTICATION_RESPONSES
 
 router = APIRouter()
