+++ backend/t4cclient/routes/notices.py
@@ -8,7 +8,7 @@ from requests import Session
 from t4cclient.core.authentication.database import verify_admin
 from t4cclient.core.authentication.jwt_bearer import JWTBearer
 from t4cclient.core.database import get_db, notices
-from t4cclient.routes.open_api_configuration import AUTHENTICATION_RESPONSES
+from t4cclient.core.oauth.responses import AUTHENTICATION_RESPONSES
 from t4cclient.schemas.notices import CreateNoticeRequest, NoticeResponse
 
 router = APIRouter()
