+++ backend/t4cclient/routes/repositories/__init__.py
@@ -19,7 +19,7 @@ from t4cclient.core.authentication.jwt_bearer import JWTBearer
 from t4cclient.core.database import get_db, repositories
 from t4cclient.core.database import users as database_users
 from t4cclient.extensions.modelsources.t4c import connection
-from t4cclient.routes.open_api_configuration import AUTHENTICATION_RESPONSES
+from t4cclient.core.oauth.responses import AUTHENTICATION_RESPONSES
 from t4cclient.schemas.repositories import (
     GetRepositoryUserResponse,
     PostRepositoryRequest,
