+++ backend/t4cclient/routes/sync.py
@@ -9,7 +9,7 @@ from requests import Session
 from t4cclient.core.authentication import database as database_auth
 from t4cclient.core.authentication.jwt_bearer import JWTBearer
 from t4cclient.core.database import get_db, repositories
-from t4cclient.routes.open_api_configuration import AUTHENTICATION_RESPONSES
+from t4cclient.core.oauth.responses import AUTHENTICATION_RESPONSES
 
 router = APIRouter()
